MS/MS peaks matching

MS/MS peaks matching

peak_matching(
  library_list = NULL,
  library_matched_features = NULL,
  RI_sample = NULL,
  ppm_err = 20,
  intensity = 20,
  RI_diff_type = c("rel", "abs")
)

Arguments

library_list

an object structured as the element "lib_precursor" of the list returned by select_library()

library_matched_features

a list of data frame, obtained by check_RT_mass() function. it is contains a list of metabolite with desired retention time and PPM error. each metabolite is a data frame of potential candidates with specific features; Feature_ID, RT_err, RT_flag, ppm_error, mass_status and mass_flag.

RI_sample

it is generated by calc_RI() function in order to turn intensity to relative intensity and filter the spectra by deleting the low intense peaks.

ppm_err

A number with default value of 10. The maximum PPM error must be less than this value. and those above this number will be eliminated.

intensity

maximum relative intensity

RI_diff_type

type of relative intensity difference: absolute or relative

Value

A list with a data.frame of matched_peaks and matrices with scores of all-pairs comparison between peaks