Create an mRList from feature data and sample information

Import data and metadata as excel or .csv or .txt format

read_input_file(
  feature_file = NULL,
  sample_file = NULL,
  MSDial = TRUE,
  split_QC = TRUE,
  mz_col = NULL,
  rt_col = NULL,
  data_start_col = NULL,
  MS_MS_column = NULL,
  ...
)

Arguments

feature_file

tab-delimited text files with feature abundances, mz, rt and MS/MS spectra

sample_file

sample information file

MSDial

(default=TRUE) the input file is MS-Dial native format

split_QC

(default=TRUE) split input and metadata in two one for QC and the other for data

mz_col

column that contains m/z value

rt_col

column that contains rt value

data_start_col

column from which data starts

MS_MS_column

column that contains the MS/MS spectrum

...

further arguments to read.delim

Value

mRList object